// src/enums/system-messages-type.ts
export enum SystemMessagesTypesEnum {
    driverAdded = 'order_driver_added',
    orderDriverUpdated = 'order_driver_updated',
    orderDriverDeleted = 'order_driver_deleted',
    orderRequestCreated = 'order_request_created',
    orderRequestRejected = 'order_request_rejected',
    orderRequestCanceled = 'order_request_canceled',
    orderOfferCreated = 'order_offer_created',
    orderOfferAccepted = 'order_offer_accepted',
    orderOfferRejected = 'order_offer_rejected',
    orderOfferCanceled = 'order_offer_canceled',
}